Cleanup integration test workflow methods
